+++ src/textual/message.py
@@ -21,6 +21,7 @@ class Message:
 
     sender: MessageTarget
     bubble: ClassVar[bool] = False
+    verbosity: ClassVar[int] = 1
 
     def __init__(self, sender: MessageTarget) -> None:
         """
@@ -39,9 +40,10 @@ class Message:
     def __rich_repr__(self) -> rich.repr.RichReprResult:
         yield self.sender
 
-    def __init_subclass__(cls, bubble: bool = False) -> None:
+    def __init_subclass__(cls, bubble: bool = False, verbosity: int = 1) -> None:
         super().__init_subclass__()
         cls.bubble = bubble
+        cls.verbosity = verbosity
 
     def can_replace(self, message: "Message") -> bool:
         """Check if another message may supersede this one.