+++ src/textual/views/_window_view.py
@@ -20,11 +20,11 @@ class WindowView(View, layout=VerticalLayout):
         self,
         widget: RenderableType | Widget,
         *,
-        gutter: tuple[int, int] = (1, 1),
+        gutter: tuple[int, int] = (0, 1),
         name: str | None = None
     ) -> None:
         self.gutter = gutter
-        layout = VerticalLayout()
+        layout = VerticalLayout(gutter=gutter)
         self.widget = widget if isinstance(widget, Widget) else Static(widget)
         layout.add(self.widget)
         super().__init__(name=name, layout=layout)
@@ -37,6 +37,7 @@ class WindowView(View, layout=VerticalLayout):
         layout.add(self.widget)
         await self.refresh_layout()
         self.require_layout()
+        await self.emit(VirtualSizeChange(self))
 
     async def watch_virtual_size(self, size: Size) -> None:
         await self.emit(VirtualSizeChange(self))